@@ -1,6 +1,6 @@
 {
     "require": {
-        "php": "7.3.*",
+        "php": "7.4.*",
         "ext-json": "*",
         "ext-pdo": "*",
         "ext-fileinfo": "*",
