@@ -7,6 +7,7 @@ namespace HomeCEU\DTS\Api\Template;
 use HomeCEU\DTS\Persistence\CompiledTemplatePersistence;
 use HomeCEU\DTS\Persistence\TemplatePersistence;
 use HomeCEU\DTS\Repository\TemplateRepository;
+use HomeCEU\DTS\UseCase\AddTemplate as AddTemplateUseCase;
 use HomeCEU\DTS\UseCase\AddTemplateRequest;
 use HomeCEU\DTS\UseCase\InvalidAddTemplateRequestException;
 use Psr\Container\ContainerInterface;
@@ -14,7 +15,7 @@ use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
 class AddTemplate {
-  private $useCase;
+  private AddTemplateUseCase $useCase;
 
   public function __construct(ContainerInterface $container) {
     $conn = $container->dbConnection;
@@ -23,7 +24,7 @@ class AddTemplate {
         new TemplatePersistence($conn),
         new CompiledTemplatePersistence($conn)
     );
-    $this->useCase = new \HomeCEU\DTS\UseCase\AddTemplate($repo);
+    $this->useCase = new AddTemplateUseCase($repo);
   }
 
   public function __invoke(Request $request, Response $response) {
