@@ -7,11 +7,11 @@ namespace HomeCEU\DTS\Entity;
 use HomeCEU\DTS\AbstractEntity;
 
 class DocData extends AbstractEntity {
-  public $dataId;
-  public $docType;
-  public $dataKey;
-  public $createdAt;
-  public $data;
+  public string $dataId;
+  public string $docType;
+  public string $dataKey;
+  public \DateTime $createdAt;
+  public array $data;
 
   protected static function keys(): array {
     return [
