@@ -3,18 +3,16 @@
 namespace HomeCEU\DTS\Entity;
 
 
-use DateTime;
 use HomeCEU\DTS\AbstractEntity;
 
 class Template extends AbstractEntity {
-  public $templateId;
-  public $templateKey;
-  public $docType;
-  public $name;
-  public $author;
-  /** @var DateTime */
-  public $createdAt;
-  public $body;
+  public string $templateId;
+  public string $templateKey;
+  public string $docType;
+  public ?string $name = '';
+  public string $author;
+  public \DateTime $createdAt;
+  public string $body;
 
   protected static function keys(): array {
     return [
