@@ -8,17 +8,15 @@ use DateTime;
 use HomeCEU\DTS\Entity\CompiledTemplate;
 use HomeCEU\DTS\Entity\Template;
 use HomeCEU\DTS\Persistence;
+use HomeCEU\DTS\Persistence\TemplatePersistence;
 use HomeCEU\DTS\Render\Image;
 use HomeCEU\DTS\Render\Partial;
 use Nette\Database\ForeignKeyConstraintViolationException;
 
 class TemplateRepository {
-  /** @var Persistence\TemplatePersistence */
-  private $persistence;
-
-  /** @var RepoHelper */
-  private $repoHelper;
-  private $compiledTemplatePersistence;
+  private Persistence $persistence;
+  private Persistence $compiledTemplatePersistence;
+  private RepoHelper $repoHelper;
 
   public function __construct(
       Persistence $persistence,
@@ -31,6 +29,7 @@ class TemplateRepository {
   }
 
   public function createNewTemplate(string $docType, string $key, string $author, string $body): Template {
+    $x = 1;
     return Template::fromState([
         'templateId' => $this->persistence->generateId(),
         'docType' => $docType,
